Added root dictionary
Included root dictionary for the tools which need to output root
files. Currently only Gater uses the root dictionary for outputting
its root tree. The reason for this change was to fix a problem with
Gater output files not being read correctly and causing the program
to crash.

Updated the Makefile to generate the new dictionaries.