Added Mike's PSD
Added Mike Febbraro's liquid scintillator PSD analyzer and
updated it to work with the current Pixie16 data structure.
Analysis.cpp and Analysis.h were created using the original
Analysis.h from the Notre Dame version of pixie scan.

Also added code that finds the leading edge, the maximum, the
minimum, and the 80% cfd bin automatically so that the values
for the integration stops no longer need to be hard coded.

Added symbolic links to point to the raw and ROOT data directories

Added a .gitignore file