Added Averager.cpp
Added a new source file named Averager.cpp which scans a number
of pulses in a branch and averages them together. The resulting
averaged pulse is then drawn to the screen.

Modified Test.cpp to now use time shift files when copying
TOFs to the output tree. If 'none' is passed as the argument
for the filename of the shift file, then the time shift is
not used.

Updated Stitcher.cpp to use the new Pixie16 data structure.
Stitcher is now able to stitch the new version of Pixie16
.root files directly.