Added DiffTime & Qcalc tools
Added two new tools named DiffTime and Qcalc.

DiffTime prints the difference between two times input from
the command line. The input and output for this program has
the format hh:mm:ss.

Qcalc calculates various beam parameters for the FN tandem
accelerator at Notre Dame. The parameters which may be
calculated are Q, Ibeam, and run time. Q is the value
from the integrator scaler. Ibeam is the beam current read
from the TwinSol FC. Run time is the length of time the
beam is run to produce the values for Q and Ibeam.